backend/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-multipart==0.0.6
requests==2.31.0
PyMuPDF==1.23.8
python-docx==1.1.0
sentence-transformers==2.2.2
transformers==4.35.0
torch==2.1.0
numpy==1.24.3
scikit-learn==1.3.2
beautifulsoup4==4.12.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4